requirements.txt
asgiref==3.4.1
Django==3.2.8
django-cors-headers==3.10.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
psycopg2==2.9.1
pytz==2021.3
sqlparse==0.4.2
